Removed PresentationCore, because it doesn't exist in Mono. Replaced Vector3D with Vector. Changed some values from double to float, because our Vector contains float values only.

+++ 01 - Code/LIFE/Examples/PedestrianModel/Config.cs
@@ -9,9 +9,9 @@ namespace PedestrianModel
 {
     class Config
     {
-        public static readonly double lengthOfTimestepsInMilliseconds = 1000d / 15d;     // 66,66666... ms, 15 frames per second
+        public static readonly float lengthOfTimestepsInMilliseconds = 1000 / 15f;     // 66,66666... ms, 15 frames per second
         public static readonly bool walkLoops = false; // if reached target, teleport to start position and start again
         public static readonly TargetListType targetListType = TargetListType.Sequential; // how to process the target positions
-        public static readonly double targetReachedDistance = 0.25; // at what distance I have reached the target position
+        public static readonly float targetReachedDistance = 0.25f; // at what distance I have reached the target position
     }
 }
